Refactored to single class with IsVertical property
