+++ src/AlohaKit/Controls/ProgressBar/Drawable/BaseProgressBarDrawable.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AlohaKit.Controls.ProgressBar
+namespace AlohaKit.Controls
 {
     public abstract class BaseProgressBarDrawable : IDrawable
     {
