+++ src/AlohaKit/Controls/ProgressBar/Drawable/HorizontalProgressBarDrawable.cs
@@ -1,9 +1,9 @@
-﻿using AlohaKit.Controls.ProgressBar;
-
+﻿
 namespace AlohaKit.Controls
 {
-    public class VerticalProgressBarDrawable : BaseProgressBarDrawable
+    public class HorizontalProgressBarDrawable : BaseProgressBarDrawable
     {
+
         public override void DrawChart(ICanvas canvas, RectF dirtyRect)
         {
             canvas.Antialias = true;
@@ -19,11 +19,12 @@ namespace AlohaKit.Controls
 
             canvas.SetFillPaint(BackgroundPaint, dirtyRect);
 
-            var x = (float)(dirtyRect.Width / 2);
-            var y = dirtyRect.Y;
+            var x = dirtyRect.X;
+            var y = (float)(dirtyRect.Height / 2);
 
             var width = dirtyRect.Width;
             var height = dirtyRect.Height;
+
 
             if (Style == ProgressBarStyle.Square)
                 canvas.FillRectangle(x, y, width, height);
@@ -39,16 +40,16 @@ namespace AlohaKit.Controls
 
             canvas.SetFillPaint(ProgressPaint, dirtyRect);
 
-            var x = (float)(dirtyRect.Width / 2);
-            var y = dirtyRect.Y + dirtyRect.Height;
+            var x = dirtyRect.X;
+            var y = (float)(dirtyRect.Height / 2);
 
             var width = dirtyRect.Width;
             var height = dirtyRect.Height;
 
             if (Style == ProgressBarStyle.Square)
-                canvas.FillRectangle(x, y, width, -(float)(height * Progress));
+                canvas.FillRectangle(x, y, (float)(width * Progress), height);
             else
-                canvas.FillRoundedRectangle(x, y, width, -(float)(height * Progress), CornerRadius);
+                canvas.FillRoundedRectangle(x, y, (float)(width * Progress), height, CornerRadius);
 
             canvas.RestoreState();
         }
