--- src/AlohaKit/Controls/ProgressBar/HorizontalProgressBar.cs
@@ -1,16 +1,0 @@
-﻿using AlohaKit.Controls.ProgressBar;
-
-namespace AlohaKit.Controls
-{
-    public class HorizontalProgressBar : BaseProgressBar
-    {
-        public HorizontalProgressBar()
-        {
-            HeightRequest = 18;
-            WidthRequest = 120;
-
-            Drawable = ProgressBarDrawable = new HorizontalProgressBarDrawable();
-        }
-
-    }
-}
