+++ src/AlohaKit/Controls/ProgressBar/ProgressBar.cs
@@ -4,27 +4,59 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AlohaKit.Controls.ProgressBar
+namespace AlohaKit.Controls
 {
-    public abstract class BaseProgressBar : GraphicsView
+    public class ProgressBar : GraphicsView
     {
         protected BaseProgressBarDrawable ProgressBarDrawable { get; set; }
 
-        public static readonly BindableProperty EasingProperty = BindableProperty.Create(nameof(Easing), typeof(Easing), typeof(BaseProgressBar), Easing.BounceOut);
+        public ProgressBar()
+        {
+            if (IsVertical)
+            {
+                Drawable = ProgressBarDrawable = new VerticalProgressBarDrawable();
+            }
+        }
+
+        public static readonly BindableProperty IsVerticalProperty = BindableProperty.Create(nameof(IsVertical), typeof(bool), typeof(ProgressBar), true,
+        propertyChanged: (bindableObject, oldValue, newValue) =>
+        {
+            if (newValue != null && bindableObject is ProgressBar progressBar)
+            {
+                if ((bool)newValue)
+                {
+                    progressBar.Drawable = progressBar.ProgressBarDrawable = new VerticalProgressBarDrawable();
+                }
+                else
+                {
+                    progressBar.Drawable = progressBar.ProgressBarDrawable = new HorizontalProgressBarDrawable();
+                }
+
+                progressBar.Invalidate();
+            }
+        });
+
+        public bool IsVertical
+        {
+            get => (bool)GetValue(IsVerticalProperty);
+            set => SetValue(IsVerticalProperty, value);
+        }
+
+        public static readonly BindableProperty EasingProperty = BindableProperty.Create(nameof(Easing), typeof(Easing), typeof(ProgressBar), Easing.BounceOut);
         public Easing Easing
         {
             get => (Easing)GetValue(EasingProperty);
             set => SetValue(EasingProperty, value);
         }
 
-        public static readonly BindableProperty EasingIntervalProperty = BindableProperty.Create(nameof(EasingInterval), typeof(int), typeof(BaseProgressBar), 1000);
+        public static readonly BindableProperty EasingIntervalProperty = BindableProperty.Create(nameof(EasingInterval), typeof(int), typeof(ProgressBar), 1000);
         public int EasingInterval
         {
             get => (int)GetValue(EasingIntervalProperty);
             set => SetValue(EasingIntervalProperty, value);
         }
 
-        public static readonly BindableProperty EnableAnimationsProperty = BindableProperty.Create(nameof(EnableAnimations), typeof(bool), typeof(BaseProgressBar), true);
+        public static readonly BindableProperty EnableAnimationsProperty = BindableProperty.Create(nameof(EnableAnimations), typeof(bool), typeof(ProgressBar), true);
         public bool EnableAnimations
         {
             get => (bool)GetValue(EnableAnimationsProperty);
@@ -33,10 +65,10 @@ namespace AlohaKit.Controls.ProgressBar
 
 
         public static readonly BindableProperty RoundCornersProperty =
-        BindableProperty.Create(nameof(RoundCorners), typeof(bool), typeof(BaseProgressBar), false,
+        BindableProperty.Create(nameof(RoundCorners), typeof(bool), typeof(ProgressBar), false,
         propertyChanged: (bindableObject, oldValue, newValue) =>
         {
-            if (newValue != null && bindableObject is BaseProgressBar progressBar)
+            if (newValue != null && bindableObject is ProgressBar progressBar)
             {
                 progressBar.UpdateRoundCorners();
             }
@@ -49,10 +81,10 @@ namespace AlohaKit.Controls.ProgressBar
         }
 
         public static readonly BindableProperty CornerRadiusProperty =
-    BindableProperty.Create(nameof(CornerRadius), typeof(float), typeof(BaseProgressBar), 6f,
+    BindableProperty.Create(nameof(CornerRadius), typeof(float), typeof(ProgressBar), 6f,
         propertyChanged: (bindableObject, oldValue, newValue) =>
         {
-            if (newValue != null && bindableObject is BaseProgressBar progressBar)
+            if (newValue != null && bindableObject is ProgressBar progressBar)
             {
                 progressBar.UpdateCornerRadius();
             }
@@ -65,10 +97,10 @@ namespace AlohaKit.Controls.ProgressBar
         }
 
         public static readonly BindableProperty ProgressBrushProperty =
-            BindableProperty.Create(nameof(ProgressBrush), typeof(Brush), typeof(BaseProgressBar), new SolidColorBrush(Colors.Blue),
+            BindableProperty.Create(nameof(ProgressBrush), typeof(Brush), typeof(ProgressBar), new SolidColorBrush(Colors.Blue),
                 propertyChanged: (bindableObject, oldValue, newValue) =>
                 {
-                    if (newValue != null && bindableObject is BaseProgressBar progressBar)
+                    if (newValue != null && bindableObject is ProgressBar progressBar)
                     {
                         progressBar.UpdateBrush();
                     }
@@ -81,10 +113,10 @@ namespace AlohaKit.Controls.ProgressBar
         }
 
         public static readonly BindableProperty ProgressProperty =
-            BindableProperty.Create(nameof(Progress), typeof(double), typeof(BaseProgressBar), 0d,
+            BindableProperty.Create(nameof(Progress), typeof(double), typeof(ProgressBar), 0d,
                 propertyChanged: (bindableObject, oldValue, newValue) =>
                 {
-                    if (newValue != null && bindableObject is BaseProgressBar progressBar)
+                    if (newValue != null && bindableObject is ProgressBar progressBar)
                     {
                         progressBar.UpdateProgress();
                     }
