+++ src/AlohaKit/Controls/ProgressBar/ProgressBarStyle.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AlohaKit.Controls.ProgressBar
+namespace AlohaKit.Controls
 {
     public enum ProgressBarStyle
     {
