--- src/AlohaKit/Controls/ProgressBar/VerticalProgressBar.cs
@@ -1,15 +1,0 @@
-﻿using AlohaKit.Controls.ProgressBar;
-
-namespace AlohaKit.Controls
-{
-    public class VerticalProgressBar : BaseProgressBar
-    {
-        public VerticalProgressBar()
-        {
-            HeightRequest = 120;
-            WidthRequest = 18;
-
-            Drawable = ProgressBarDrawable = new VerticalProgressBarDrawable();
-        }
-    }
-}
